==> code/00_EDA.Rmd <==
---
title: "KELPER2"
subtitle: "Dataset EDA"
author: "Tim Szewczyk"
output: 
  html_document: 
    number_sections: yes
    theme: spacelab
    anchor_sections: yes
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_knit$set(root.dir=rprojroot::find_rstudio_root_file())
knitr::opts_chunk$set(echo=TRUE, message=F, warning=F)
library(tidyverse); library(glue); library(lubridate)
theme_set(theme_bw())

data.dir <- "data/chile/"
```

This document is for preliminary exploration of the data from the Subtidal Ecology lab at Pontificia Universidad. See [data-collection.html](data-collection.html) for details on the collection.


# Allometry

Start with the simplest. Holdfast diameter will be the primary metric for tracking plant size and will be used to estimate everything else, including biomass and demographic rates.

```{r allom_loadData}
morph.df <- list(
  IFOP_morph=read_csv(glue("{data.dir}IFOP_Kelp_Morphology.csv"),
                      na=c("NaN", "")) %>% 
    select(c(1:9,11:13,15:19)) %>%
    filter(Group=="Macroalga" &
             Species=="Lessonia trabeculata" &
             Abundance > 0) %>% 
    mutate(Date=ymd(paste(Year, Month, Day, sep="-")),
           Site=factor(Site),
           Management=factor(Management),
           Frondosity=factor(Frondosity),
           source="IFOP_morph") %>%
    rename(LengthTot=TL),
  IFOP_allom=read_csv(glue("{data.dir}IFOP_allometry.csv")) %>% 
    select(-16) %>%
    mutate(Date=ymd(paste(Year, Month, Day, sep="-")),
           Site=factor(Site),
           Management=factor(Management),
           Weight=1e3*Weight,
           source="IFOP_allom") %>%
    rename(LengthTot=TL,
           WeightTot=Weight),
  NERC_bl=read_csv(glue("{data.dir}Quads_baseline.csv")) %>% 
    select(-c(1,3:4,10:12)) %>%
    filter(Species=="Lessonia trabeculata") %>%
    mutate(Site=factor(Site),
           Zone=factor(Zone),
           Management=factor(Management),
           Upwelling=factor(Upwelling),
           Frondosity=factor(Frondosity),
           source="NERC_bl") %>%
    rename(Holdfast=Holdfast.Diameter,
           LengthTot=Total.Length),
  NERC_ph=read_csv(glue("{data.dir}Patches_Morphology.csv")) %>%
    mutate(Site=factor(Site),
           Zone=factor(Zone),
           Management=factor(Management),
           Upwelling=factor(Upwelling),
           source="NERC_ph") %>%
    rename(Holdfast=Holdfast.Diameter,
           LengthTot=Total.Length),
  Cata_ph=read_csv(glue("{data.dir}Morfo_DW_C.csv")) %>%
    select(-c(1,11,20:30)) %>%
    mutate(Site=factor(Site),
         Zone=factor(Zone),
         Management=factor(Management)) %>%
    rename(Holdfast=Diam_Holdfast,
           Height=High_Holdfast,
           LengthTot=LT,
           WeightTot=TW_sum,
           Weight=Weight_Str,
           Length=LT_stipes,
           Dichot=Dichot_stipes,
           Diam=Diam_stipes) %>%
    pivot_wider(names_from="Structure", values_from=c(8:17)) %>%
    select(!where(~all(is.na(.x)))) %>%
    select(-starts_with("Structure_")) %>%
    mutate(source="Cata_ph") %>%
    rename(Stipes=Stipes_Stipes)
) %>%
  do.call('bind_rows', .)
```

```{r allom_summaries}
glimpse(morph.df)
summary(morph.df)
```

```{r morph_univariate}
ggplot(morph.df, aes(Holdfast)) + geom_histogram() + 
  facet_wrap(~source, scales="free_y")
ggplot(morph.df, aes(Holdfast, colour=source)) + geom_density()

ggplot(morph.df, aes(LengthTot)) + geom_histogram() + 
  facet_wrap(~source, scales="free_y")
ggplot(morph.df, aes(LengthTot, colour=source)) + geom_density()

ggplot(morph.df, aes(WeightTot)) + geom_histogram() + 
  facet_wrap(~source, scales="free_y")
ggplot(morph.df, aes(WeightTot, colour=source)) + geom_density()

ggplot(morph.df, aes(Frondosity)) + geom_bar() + 
  facet_wrap(~source, scales="free_y")

ggplot(morph.df, aes(factor(Reproductive))) + geom_bar() + 
  facet_wrap(~source, scales="free_y")

```


```{r allom_plots}
ggplot(morph.df, aes(Holdfast, Stipes, colour=source)) + 
  geom_point(alpha=0.5, shape=1) + 
  stat_smooth(method="lm", se=F, size=0.5, linetype=3, aes(group=Site)) +
  stat_smooth(method="lm", se=F, size=0.75, aes(group=source)) +
  scale_x_continuous("Holdfast diameter", trans="log1p") +
  scale_y_continuous(trans="log1p")

ggplot(morph.df, aes(Holdfast, LengthTot, colour=source)) + 
  geom_point(alpha=0.5, shape=1) + 
  stat_smooth(method="lm", se=F, size=0.5, linetype=3, aes(group=Site),
              formula=y~x+I(x^2)) +
  stat_smooth(method="lm", se=F, size=0.75, aes(group=source),
              formula=y~x+I(x^2)) +
  scale_x_continuous("Holdfast diameter", trans="log1p") +
  scale_y_continuous(trans="log1p")

ggplot(morph.df, aes(Holdfast, LengthTot, colour=Management)) + 
  geom_point(alpha=0.5, shape=1) + 
  stat_smooth(method="lm", se=F, size=0.5, linetype=3, aes(group=Site),
              formula=y~x+I(x^2)) +
  stat_smooth(method="lm", se=T, size=0.75, aes(group=Management),
              formula=y~x+I(x^2)) +
  scale_x_continuous("Holdfast diameter", trans="log1p") +
  scale_y_continuous(trans="log1p")

ggplot(morph.df, aes(Holdfast, WeightTot, colour=source)) + 
  geom_point(alpha=0.5, shape=1) + 
  stat_smooth(method="lm", se=F, size=0.5, linetype=3, aes(group=Site),
              formula=y~x) +
  stat_smooth(method="lm", se=F, size=0.75, aes(group=source),
              formula=y~x) +
  scale_x_continuous("Holdfast diameter", trans="log1p") +
  scale_y_continuous(trans="log1p")

ggplot(morph.df, aes(Holdfast, WeightTot, colour=Management)) + 
  geom_point(alpha=0.5, shape=1) + 
  stat_smooth(method="lm", se=F, size=0.5, linetype=3, aes(group=Site),
              formula=y~x) +
  stat_smooth(method="lm", se=T, size=0.75, aes(group=Management),
              formula=y~x) +
  scale_x_continuous("Holdfast diameter", trans="log1p") +
  scale_y_continuous(trans="log1p")

ggplot(morph.df, aes(LengthTot, WeightTot, colour=source)) + 
  geom_point(alpha=0.5, shape=1) + 
  stat_smooth(method="lm", se=F, size=0.5, linetype=3, aes(group=Site)) +
  stat_smooth(method="lm", se=F, size=0.75, aes(group=source)) +
  scale_x_continuous(trans="log1p") +
  scale_y_continuous(trans="log1p")

ggplot(morph.df, aes(Holdfast, Weight_Holdfast, colour=source)) + 
  geom_point(alpha=0.5, shape=1) + 
  stat_smooth(method="lm", se=F, size=0.5, linetype=3, aes(group=Site)) +
  stat_smooth(method="lm", se=F, size=0.75, aes(group=source)) +
  scale_x_continuous("Holdfast diameter", trans="log1p") +
  scale_y_continuous(trans="log1p")
ggplot(morph.df, aes(Holdfast, Weight_Stipes, colour=source)) + 
  geom_point(alpha=0.5, shape=1) + 
  stat_smooth(method="lm", se=F, size=0.5, linetype=3, aes(group=Site)) +
  stat_smooth(method="lm", se=F, size=0.75, aes(group=source)) +
  scale_x_continuous("Holdfast diameter", trans="log1p") +
  scale_y_continuous(trans="log1p")
ggplot(morph.df, aes(Holdfast, Weight_Blades, colour=source)) + 
  geom_point(alpha=0.5, shape=1) + 
  stat_smooth(method="lm", se=F, size=0.5, linetype=3, aes(group=Site)) +
  stat_smooth(method="lm", se=F, size=0.75, aes(group=source)) +
  scale_x_continuous("Holdfast diameter", trans="log1p") +
  scale_y_continuous(trans="log1p")

# Larger holdfast = less blade-y by weight, more holdfast-y & stipe-y
ggplot(morph.df, aes(Holdfast, Weight_Holdfast/WeightTot, colour=source)) + 
  geom_point(alpha=0.5, shape=1) + 
  stat_smooth(method="lm", se=F, size=0.5, linetype=3, aes(group=Site)) +
  stat_smooth(method="lm", se=F, size=0.75, aes(group=source)) +
  scale_x_continuous("Holdfast diameter", trans="log1p") +
  scale_y_continuous("Holdfast weight / Total", limits=c(0,1))
ggplot(morph.df, aes(Holdfast, Weight_Stipes/WeightTot, colour=source)) + 
  geom_point(alpha=0.5, shape=1) + 
  stat_smooth(method="lm", se=F, size=0.5, linetype=3, aes(group=Site)) +
  stat_smooth(method="lm", se=F, size=0.75, aes(group=source)) +
  scale_x_continuous("Holdfast diameter", trans="log1p") +
  scale_y_continuous("Stipe weight / Total", limits=c(0,1))
ggplot(morph.df, aes(Holdfast, Weight_Blades/WeightTot, colour=source)) + 
  geom_point(alpha=0.5, shape=1) + 
  stat_smooth(method="lm", se=F, size=0.5, linetype=3, aes(group=Site)) +
  stat_smooth(method="lm", se=F, size=0.75, aes(group=source)) +
  scale_x_continuous("Holdfast diameter", trans="log1p") +
  scale_y_continuous("Blade weight / Total", limits=c(0,1))

ggplot(morph.df, aes(Holdfast, Height_Holdfast, colour=source)) + 
  geom_point(alpha=0.5, shape=1) + 
  stat_smooth(method="lm", se=F, size=0.25, aes(group=Site)) +
  scale_x_log10() + scale_y_log10()

ggplot(morph.df, aes(Holdfast, colour=factor(Frondosity))) + 
  geom_density() + scale_colour_viridis_d() +
  scale_x_log10() 

# Lots of NAs -- Cata's thesis should have this too
ggplot(morph.df, aes(Holdfast, colour=factor(Reproductive))) + 
  geom_density() + 
  scale_x_log10() 
```


# Density
```{r density_loadData}
dens.df <- bind_rows(
  read_csv(glue("{data.dir}IFOP_Kelp_density.csv"), na=c("NaN", "")) %>% 
    filter(Species=="Lessonia trabeculata") %>%
    mutate(source="IFOP",
           Date=ymd(paste(Year, Month, Day, sep="-")),
           Stage=if_else(Stage=="Adulto", "Adult", "Juvenile"),
           Density_m2=Abundance/20) %>%
    select(source, Management, Date, Site, Transect, Station, Depth, Stage, Density_m2) %>%
    group_by(source, Management, Date, Site, Transect, Station, Depth, Stage) %>%
    summarise(Density_m2=mean(Density_m2)) %>%
    ungroup,
  read_csv(glue("{data.dir}Kelp_transect.csv")) %>%
    mutate(source="NERC_bl",
           Density_m2=Density) %>%
    select(source, Management, Zone, Date, MAE, Site, Transect, Station, Depth, Stage, Density_m2),
  read_csv(glue("{data.dir}Patches_quad.csv")) %>%
    filter(Species=="Lessonia trabeculata") %>%
    mutate(source="NERC_ph",
           Density_m2=Density) %>%
    select(source, Management, Zone, Date, MAE, Site, Patch, Quad, Stage, Density_m2)
)

```

```{r density_plots}
dens.df %>% 
  pivot_wider(names_from=Stage, values_from=Density_m2) %>% 
  ggplot(aes(Adult, Juvenile)) + 
  geom_point(shape=1, alpha=0.25) + 
  geom_line(aes(group=Site), stat="smooth", method="lm", se=F, colour="blue") + 
  scale_y_continuous(trans="log1p") +
  scale_x_continuous(trans="log1p") +
  facet_grid(Management~source)

dens.df %>% filter(source=="NERC_bl") %>%
  ggplot(aes(Density_m2, fill=Management)) + geom_histogram(binwidth=0.5, position="dodge") + 
  facet_wrap(~Stage)

```







# Growth & Survival dataset
```{r growthSurv_loadData}
Growth_Surv <- read_csv(glue("{data.dir}Growth&Survival.csv")) %>%
  mutate(Site=factor(Site),
         Management=factor(Management),
         Plant=factor(Plant),
         alive=as.numeric(Holdfast > 0),
         Date_Harvest=Date_harvest_dmy %>%
           str_replace("Dic", "01-12") %>%
           str_replace("Nov", "01-11") %>%
           str_replace("Ene", "01-01") %>%
           str_replace("Feb", "01-02") %>%
           dmy(),
         Date_Obs=Date_monitoring %>%
           str_replace("Abr", "01-04") %>%
           str_replace("Oct", "01-10") %>%
           str_replace("May", "01-05") %>%
           dmy()) %>%
  mutate(days_since_harvest=as.numeric(Date_Obs - Date_Harvest)) %>%
  filter(!is.na(Site))

survey.i <- Growth_Surv %>%
  group_by(Site, Management, Patch, Monitoring) %>%
  summarise(Date_Harvest=min(Date_Harvest),
            Date_Obs=min(Date_Obs)) %>%
  arrange(Management, Site, Patch, Monitoring)
survey.i <- bind_rows(
  survey.i,
  survey.i %>% slice_head(n=1) %>%
    mutate(Monitoring=0,
           Date_Obs=Date_Harvest)
  ) %>%
  arrange(Management, Site, Patch, Monitoring) %>%
  mutate(days_since_harvest=as.numeric(Date_Obs - Date_Harvest),
         days_since_prevObs=as.numeric(Date_Obs - lag(Date_Obs)),
         days_since_prevObs=replace_na(days_since_prevObs, 0),
         days_to_nextObs=as.numeric(lead(Date_Obs) - Date_Obs))

Growth_Surv <- Growth_Surv %>%
  select(Site, Management, Patch, Monitoring, 
         Plant, Holdfast, Stipes, Total_Length) %>%
  left_join(., survey.i) %>%
  arrange(Site, Patch, Plant, Monitoring) %>%
  mutate(prevSurv=Monitoring-1) %>%
  arrange(Site, Patch, Plant, Monitoring) %>%
  group_by(Site, Patch, Plant) %>%
  mutate(doesSurvive=lead(Holdfast>0),
         nextHoldfast=lead(Holdfast),
         nextStipes=lead(Stipes),
         nextLength=lead(Total_Length),
         maxAge_days=cumsum(days_since_prevObs),
         maxAge_yrs=maxAge_days/365,
         surv_01=as.numeric(doesSurvive),
         logHoldfast=log(Holdfast)) %>%
  ungroup
```




# Survival
We can link probability of survival to holdfast size x Management. With the magic of Hierarchical Bayesian modelling, we can annualize the estimated survival rates if we assume that the survival probability is constant throughout the year.
```{r survival_plots}
surv.df <- Growth_Surv %>% filter(!is.na(doesSurvive))

surv.df %>% group_by(Management) %>%
  summarise(nPlants=n_distinct(paste(Site, Patch, Plant)),
            nMortalities=sum(!doesSurvive),
            nSurvivals=sum(doesSurvive))

ggplot(surv.df, aes(Holdfast, as.numeric(doesSurvive))) +  
  geom_jitter(shape=1, height=0.025, width=0.25) +
  stat_smooth(method="glm", method.args=list(family="binomial")) +
  scale_x_continuous(trans="log", breaks=c(0.1, 1, 5, 20))
ggplot(surv.df, aes(Holdfast, as.numeric(doesSurvive), colour=Management)) +  
  geom_jitter(shape=1, height=0.025, width=0.25) +
  stat_smooth(method="glm", method.args=list(family="binomial")) +
  scale_x_continuous(trans="log", breaks=c(0.1, 1, 5, 20))

# This is asking a lot of the data, but we can get reasonable looking curves
# across management types. The outcome isn't all that sensitive to priors 
# at least within reason. Here's an option expecting greater variability among
# sites, then among plants, then among patches, and some nudged slopes.
library(brms)
out.s <- brm(bf(surv_01 ~ inv_logit(pSurvYr)^maxAge_yrs,
                pSurvYr ~ logHoldfast*Management + 
                  (1+logHoldfast*Management|Site/Patch/Plant),
                nl=T),
             data=surv.df, init=0, family="bernoulli", chains=4, cores=4,
             prior=c(prior(normal(0, 1), nlpar="pSurvYr"),
                     prior(normal(-1, 1), nlpar="pSurvYr", coef="Intercept"),
                     prior(normal(1, 1), nlpar="pSurvYr", coef="ManagementTURF"),
                     prior(normal(0, 0.5), nlpar="pSurvYr", class="sd", lb=0, 
                           group="Site"),
                     prior(normal(0, 0.1), nlpar="pSurvYr", class="sd", lb=0,
                           group="Site:Patch"),
                     prior(normal(0, 0.25), nlpar="pSurvYr", class="sd", lb=0,
                           group="Site:Patch:Plant")))

fitted.df <- expand_grid(logHoldfast=seq(-2.5,3.5,length.out=100),
            Management=levels(surv.df$Management)) %>%
  mutate(maxAge_yrs=1,
         Holdfast=exp(logHoldfast),
         Site=NA, Patch=NA, Plant=NA) %>%
  bind_cols(posterior_epred(out.s, newdata=., nlpar="pSurvYr", 
                            ndraws=4000, re_formula=NA) %>%
                        t %>% as_tibble) %>%
  pivot_longer(starts_with("V"), names_to="iter", values_to="logit_p") %>%
  group_by(logHoldfast, Holdfast, Management) %>%
  summarise(logit_p_mn=mean(logit_p),
            logit_p_lo=tidybayes::qi(logit_p, 0.9)[,1],
            logit_p_hi=tidybayes::qi(logit_p, 0.9)[,2]) %>%
  mutate(across(starts_with("logit_p"), inv_logit_scaled, 
                .names="{str_sub({.col}, 7, -1)}"))
fitted.df %>%
  ggplot(aes(Holdfast, p_mn, colour=Management, fill=Management)) + 
  geom_ribbon(aes(ymin=p_lo, ymax=p_hi), alpha=0.5, colour=NA) +
  geom_line() +
  geom_jitter(data=surv.df, aes(y=surv_01),
              shape=1, height=0.025, width=0.05) +
  scale_x_continuous(trans="log", breaks=c(0.1, 1, 5, 20)) +
  labs(x="Holdfast diameter (cm)", y="Annual survival probability")

fitted.df %>%
  ggplot(aes(logHoldfast, logit_p_mn, colour=Management, fill=Management)) + 
  geom_ribbon(aes(ymin=logit_p_lo, ymax=logit_p_hi), alpha=0.5, colour=NA) +
  geom_line() +
  geom_jitter(data=surv.df, aes(y=logit_scaled(pmax(pmin(surv_01, 0.999), 0.001))),
              shape=1, height=0.025, width=0.05) +
  labs(x="log holdfast diameter (cm)", y="Annual survival probability")
```



# Growth
We can link the holdfast size at t+1 to holdfast size at t x Management. Like with mortality, we can assume a constant daily growth rate to account for differences in the number of days between surveys and calculate annualized growth rates by holdfast diameter.
```{r growth_plots}
grow.df <- Growth_Surv %>% 
  filter(doesSurvive) %>%
  mutate(annHoldfastGrowth=(nextHoldfast-Holdfast)/days_to_nextObs*365,
         annStipeGrowth=(nextStipes-Stipes)/days_to_nextObs*365,
         annLengthGrowth=(nextLength-Total_Length)/days_to_nextObs*365,
         nextHoldfastAnnual=Holdfast+annHoldfastGrowth,
         logNextHoldfastAnnual=log(nextHoldfastAnnual))

ggplot(grow.df, aes(Holdfast, Holdfast+annHoldfastGrowth)) + 
  geom_abline(colour="grey") +
  geom_point() + stat_smooth(method="lm", formula=y~x+I(x^2)) +
  scale_x_log10() + scale_y_log10()
ggplot(grow.df, aes(Holdfast, Holdfast+annHoldfastGrowth, colour=Management)) +
  geom_abline(colour="grey") +
  geom_point() + 
  stat_smooth(aes(group=Site), method="lm", formula=y~x+I(x^2), se=F, size=0.25) +
  stat_smooth(method="lm", formula=y~x+I(x^2)) +
  scale_x_log10() + scale_y_log10()

ggplot(grow.df, aes(Stipes, Stipes+annStipeGrowth)) +
  geom_abline(colour="grey") +
  geom_point() + stat_smooth(method="lm", formula=y~x+I(x^2)) 
ggplot(grow.df, aes(Stipes, Stipes+annStipeGrowth, colour=Management)) + 
  geom_abline(colour="grey") +
  geom_point() + 
  stat_smooth(aes(group=Site), method="lm", formula=y~x+I(x^2), se=F, size=0.25) +
  stat_smooth(method="lm", formula=y~x+I(x^2)) 

ggplot(grow.df, aes(Total_Length, Total_Length+annLengthGrowth)) + 
  geom_abline(colour="grey") +
  geom_point() + stat_smooth(method="lm", formula=y~x+I(x^2))
ggplot(grow.df, aes(Total_Length, Total_Length+annLengthGrowth, colour=Management)) +
  geom_abline(colour="grey") +
  geom_point() + 
  stat_smooth(aes(group=Site), method="lm", formula=y~x+I(x^2), se=F, size=0.25) +
  stat_smooth(method="lm", formula=y~x+I(x^2)) 


# log-log scale, heteroscedastic
out.g <- brm(bf(logNextHoldfastAnnual ~ logHoldfast*Management + 
                  (1+logHoldfast*Management|Site/Patch/Plant),
                sigma ~ logHoldfast),
             family="gaussian", data=grow.df, cores=4,
             prior=c(prior(normal(0, 0.1), class="sd", lb=0, group="Site"),
                     prior(normal(0, 0.1), class="sd", lb=0, group="Site:Patch"),
                     prior(normal(0, 0.1), class="sd", lb=0, group="Site:Patch:Plant"),
                     prior(normal(0, 1), class="b"),
                     prior(normal(0, 1), class="b", dpar="sigma"),
                     prior(normal(2, 1), class="Intercept"),
                     prior(normal(0, 1), class="Intercept", dpar="sigma")))

fitted.df <- expand_grid(Holdfast=seq(0.1,44,length.out=100),
                         Management=levels(surv.df$Management)) %>%
  mutate(yrs_to_next=1,
         logHoldfast=log(Holdfast),
         Site=NA, Patch=NA, Plant=NA)
preds <- posterior_epred(out.g, newdata=fitted.df, re.form=NA)
# preds <- posterior_predict(out.g, newdata=fitted.df, re.form=NA)
fitted.df <- fitted.df %>%
  mutate(pred_mn=colMeans(preds),
         pred_lo=apply(preds, 2, function(x) quantile(x, probs=0.1)),
         pred_hi=apply(preds, 2, function(x) quantile(x, probs=0.9)),
         pred_emn=colMeans(exp(preds)),
         pred_elo=apply(exp(preds), 2, function(x) quantile(x, probs=0.1)),
         pred_ehi=apply(exp(preds), 2, function(x) quantile(x, probs=0.9)))
fitted.df %>%
  ggplot(aes(logHoldfast, pred_mn, colour=Management, fill=Management)) + 
  geom_abline() +
  geom_ribbon(aes(ymin=pred_lo, ymax=pred_hi), alpha=0.5, colour=NA) +
  geom_line() +
  geom_point(data=grow.df, aes(y=log(Holdfast+annHoldfastGrowth)), shape=1) +
  labs(x="Holdfast diameter (t)", y="Holdfast diameter (t+1)")

fitted.df %>%
  ggplot(aes(Holdfast, pred_emn, colour=Management, fill=Management)) + 
  geom_abline() +
  geom_ribbon(aes(ymin=pred_elo, ymax=pred_ehi), alpha=0.5, colour=NA) +
  geom_line() +
  geom_point(data=grow.df, aes(y=Holdfast+annHoldfastGrowth), shape=1) +
  labs(x="Holdfast diameter (t)", y="Holdfast diameter (t+1)") +
  xlim(0, 50) + ylim(0, 50)


```





# Recruitment
We can estimate size distributions of recruits by management, and estimate the number of recruits as a function of the number or size of non-recruits in the plot. **TODO:** Account for  the differences in time between monitoring periods to annualize somehow.
```{r recruitment_plots}
rcrSize.df <- Growth_Surv %>% 
  select(Site, Management, Patch, Monitoring, 
         Plant, Holdfast, Stipes, Total_Length) %>%
  left_join(., survey.i) %>%
  arrange(Site, Patch, Plant, Monitoring) %>%
  mutate(prevSurv=Monitoring-1) %>%
  arrange(Site, Patch, Plant, Monitoring) %>%
  group_by(Site, Patch, Plant) %>%
  mutate(maxAge_days=cumsum(days_since_prevObs),
         maxAge_yrs=maxAge_days/365) %>%
  group_by(Site, Patch, Monitoring) %>%
  mutate(nPlants=sum(Holdfast>0),
         totHoldfast=sum(pi*(Holdfast/2)^2)) %>%
  ungroup %>%
  filter(Monitoring==1,
         Holdfast < 10)

rcrDens.df <- read_csv(glue("{data.dir}Patches_quad.csv")) %>%
  filter(Species=="Lessonia trabeculata") %>%
  mutate(Site=factor(Site),
         Zone=factor(Zone),
         Management=factor(Management),
         Upwelling=factor(Upwelling),
         Stage=factor(Stage))
ggplot(rcrDens.df, aes(MAE, Density)) + geom_point(alpha=0.5) +
  facet_grid(Stage~Management)
ggplot(rcrDens.df, aes(factor(MAE), Density)) + geom_boxplot() +
  scale_y_continuous(trans="log1p") +
  facet_grid(Stage~Management)
ggplot(rcrDens.df, aes(Density, colour=Stage)) + 
  geom_density(alpha=0.5) +
  scale_x_continuous(trans="log1p") +
  facet_grid(MAE~Management)

rcrDens.df %>% select(-...1) %>%
  pivot_wider(names_from=Stage, values_from=Density) %>%
  ggplot(aes(Adult, Juvenile)) + 
  geom_jitter(alpha=0.5, width=0.1, height=0.1, aes(colour=Management)) +
  facet_wrap(~MAE)

rcrDens.df %>% select(-...1) %>%
  pivot_wider(names_from=Stage, values_from=Density) %>%
  ggplot(aes(Adult, Juvenile)) + 
  geom_jitter(alpha=0.5, width=0.1, height=0.1, aes(colour=Management)) +
  stat_smooth(method="lm") +
  scale_y_continuous(trans="log1p")



# Some bias in measurements: preference for 0.2, 10 cm
table(rcrSize.df$Holdfast)

ggplot(rcrSize.df, aes(Holdfast)) + geom_density() + 
  ggtitle("Recruit size distr.") +
  scale_x_continuous(trans="log1p")
ggplot(rcrSize.df, aes(Holdfast, colour=Management)) + 
  geom_density(aes(group=Site), size=0.25) + 
  geom_density(size=1) +
  ggtitle("Recruit size distr.") +
  scale_x_continuous(trans="log1p")

ggplot(rcrSize.df, aes(maxAge_yrs)) + geom_histogram() +
  ggtitle("Recruit max possible age")

ggplot(rcrSize.df, aes(Holdfast)) + geom_histogram() +
  scale_x_continuous(trans="log1p") + facet_grid(Management~.)


# THIS DOES NOT WORK WITH THIS DATASET!
# TODO: These calculations but with density (patches_quad.csv)
# rcrSize.df %>% group_by(Site, Management, Patch, Monitoring, nPlants) %>%
#   summarise(nRecruits=sum(newRecruit)) %>%
#   ggplot(aes(nPlants-nRecruits, nRecruits)) + 
#   geom_jitter(width=0.2, height=0.2) +
#   stat_smooth(method="lm", formula=y~log1p(x)) +
#   scale_y_continuous(trans="log1p") +
#   xlab("Number of non-recruits")
# rcrSize.df %>% group_by(Site, Management, Patch, Monitoring, nPlants) %>%
#   summarise(nRecruits=sum(newRecruit)) %>%
#   ggplot(aes(nPlants-nRecruits, nRecruits, colour=Management)) + 
#   geom_jitter(width=0.2, height=0.2) +
#   stat_smooth(aes(group=Site), method="lm", formula=y~log1p(x), se=F, size=0.25) +
#   stat_smooth(method="lm", formula=y~log1p(x), size=1) +
#   scale_y_continuous(trans="log1p") +
#   xlab("Number of non-recruits")
# 
# rcrSize.df %>% group_by(Site, Management, Patch, Monitoring, totHoldfast) %>%
#   summarise(nRecruits=sum(newRecruit),
#             HoldfastRcr=sum(newRecruit*pi*(Holdfast/2)^2)) %>%
#   ggplot(aes(totHoldfast-HoldfastRcr, nRecruits)) + 
#   geom_jitter(width=0.2, height=0.2) +
#   stat_smooth(method="lm", formula=y~log1p(x)) +
#   scale_y_continuous(trans="log1p", limits=c(0,NA)) +
#   xlab("Total holdfast area of non-recruits")
# rcrSize.df %>% group_by(Site, Management, Patch, Monitoring, totHoldfast) %>%
#   summarise(nRecruits=sum(newRecruit),
#             HoldfastRcr=sum(newRecruit*pi*(Holdfast/2)^2)) %>%
#   ggplot(aes(totHoldfast-HoldfastRcr, nRecruits, colour=Management)) + geom_jitter() +
#   stat_smooth(aes(group=Site), method="lm", formula=y~log1p(x), se=F, size=0.25) +
#   stat_smooth(method="lm", formula=y~log1p(x), size=1) +
#   scale_y_continuous(trans="log1p", limits=c(0,NA)) +
#   xlab("Total holdfast area of non-recruits")

# The recruit size distribution seems to change with time-since-harvest.
# Mainly small recruits at the start, then larger recruits at TURF sites as 
# populations become established... These should only be *new* recruits since
# the last survey. Differences in thinning / establishment by management? OA
# tend to have more and smaller recruits after the initial pulse. 

# How does this interact with higher survival rates and faster growth rates in TURF?
# Lower establishment, but better survival and growth once established?
# ggplot(rcrSize.df, aes(days_since_harvest, Holdfast)) + 
#   geom_jitter(width=10, height=0.1) + 
#   stat_smooth(span=1.5) +
#   ggtitle("Recruit size vs. time since clearance") +
#   scale_y_continuous(trans="log1p")
# ggplot(rcrSize.df, aes(days_since_harvest, Holdfast, colour=Management)) + 
#   geom_jitter(width=10, height=0.1) + 
#   stat_smooth(aes(group=Site), span=1.5, se=F, size=0.25) +
#   stat_smooth(span=1.5) +
#   ggtitle("Recruit size vs. time since clearance") +
#   scale_y_continuous(trans="log1p")
# 
# ggplot(rcrSize.df, aes(Holdfast, colour=factor(Monitoring))) + 
#   geom_density(aes(group=paste(Site, Monitoring)), adjust=1.5) + 
#   ggtitle("Recruit size vs. time since clearance") +
#   scale_y_continuous(trans="log1p") +
#   facet_wrap(~Management)
# 
# rcrSize.df %>% group_by(Site, Management, Patch, Monitoring, days_since_harvest) %>%
#   summarise(nRecruits=sum(newRecruit)) %>%
#   ggplot(aes(days_since_harvest, nRecruits)) + 
#   geom_jitter(width=10, height=0.1) + 
#   stat_smooth(span=1.5) +
#   ggtitle("Number of recruits vs. time since clearance") +
#   scale_y_continuous(trans="log1p")
# rcrSize.df %>% group_by(Site, Management, Patch, Monitoring, days_since_harvest) %>%
#   summarise(nRecruits=sum(newRecruit)) %>%
#   ggplot(aes(days_since_harvest, nRecruits, colour=Management)) + 
#   geom_jitter(width=10, height=0.1) + 
#   stat_smooth(aes(group=Site), span=1.5, se=F, size=0.25) +
#   stat_smooth(span=1.5) +
#   ggtitle("Number of recruits vs. time since clearance") +
#   scale_y_continuous(trans="log1p")
```





# Reproduction
We can fit a nice curve predicting the probability of reproductive blades given holdfast diameter, but do we have any data linking reproductive blades to recruits? That is, what we need is $F(z',z) \sim p_{repro}(z) f_{spores}(z) p_{estab} f_{rcrSize}(z')$. We have an estimate for $p_{repro}(z)$, which is the probability of having reproductive blades given a holdfast size $z$. We have part of $f_{rcrSize}(z')$, which is the expected recruit size distribution at time $t+1$. What we're missing is how many spores might be produced by each reproductive plant, and how likely those spores are to establish. 

We could potentially lump several processes and calculate directly the number of recruits from the data above based on a presumed maximum established recruits, with the actual as a function of $N$.
```{r}
morph.df %>%
  filter(!is.na(Reproductive)) %>% 
  droplevels() %>%
  ggplot(aes(Holdfast, Reproductive)) + 
  geom_jitter(shape=1, height=0.025, width=0.5) +
  stat_smooth(method="glm", method.args=list(family="binomial")) +
  scale_x_continuous(trans="log1p")
morph.df %>%
  filter(!is.na(Reproductive)) %>% 
  droplevels() %>%
  ggplot(aes(Holdfast, Reproductive, colour=Management)) + 
  geom_jitter(shape=1, height=0.025, width=0.5) +
  stat_smooth(method="glm", method.args=list(family="binomial")) +
  scale_x_continuous(trans="log1p")
```
